Wire frame workign with new Rdata File.
Updated wireframing to use the structure of the new Rdata file.  I split the old file input reactive into an observe event to load in the file and a reactive to incorperate it.

Now you can select a unit number from the "Unit Number" list on the side bar and the tables and inputs updated.